Enable certain field declarations and improve error messges
This PR allows the user to declare a field like this:

    var myField: someType = someExpr;

where `someType` disallows 0-argument initialization, ex:

    test/types/records/init/field-typed-dfltexpr.chpl

This is done by modifying `proc _createFieldDefault` in ChapelBase.chpl:

    var x: t;  x = initVal;  // before: init() plus assignment
    var x: t = initVal;      // now: init(initVal)

with the side effect of eliminating default initialization of `x` and
deinitialization of `initVal`.

The above change is what is important in this PR. However, by itself,
it wrecks havoc in the error messages that are generated ex. when
`t` is owned or non-nilable and `initVal` is unmanaged or nilable.

The compiler changes restore order and improve the error messages
along the way.

While there:
* remove nilability in a tuple type in `studies/amr`
* slightly simplify `unmanaged-concrete.chpl`