Add intents and lifetimes to several list routines (#13795)
Add intents and lifetimes to several list routines (#13795)

This is a small PR that gives the list routines `append` and `insert`
the in intent. It also adds lifetime checks to `append`, `insert`,
and `extend`. These changes will enable the `list` type to behave
properly when its element type is an owned class.

This PR also fixes the test `stridedAssignDR.chpl` until a solution
to #13808 is found.

Testing:

- [x] `ALL` on `linux64` when `CHPL_COMM=none`

Trivial and not reviewed.