Add Maps module implementing the map data type
Implement the map interface from issue #13624 and chapel-private/#366.

This includes procedures to add key-value pairs, set values that are already
in the map, remove keys, clear the map, iterate over keys, key-value pairs, or
values, and make arrays from the map.

It also includes the map-as-set operators that were previously array-as-set
operators. It does not yet deprecate the array-as-set operators.

Added tests for all of the new operations.