Do some housecleaning in ChapelStandard.chpl
This is an attempt to reduce the number of modules `use`d in
ChapelStandard to focus on (a) those required to implement the
language and (b) the standard modules that we've decided to
`use` by default.  Probably more could be done here, but this
is a step in the right direction.