Task intents to operate on fields of 'this'
Resolves #13577.

This is the counterpart for task intents to #13479, which implemented
the proposals in #13297 for forall-intents. Which is that ...

When a task construct, i.e. `coforall`, `cobegin`, or `begin`,
accesses a field of `this` that is a record, this field is treated
as a variable passed into the task function by default intent.
As opposed to passing `this` into the task function by default intent and
accessing a field off of the corresponding task-private/shadow variable.

The implementation reuses much of the code from #13479. Including
the general approach of introducing a `ref` variable aliasing
the record's field and passing that variable into the task function.

I chose to place the new code in implementForallIntents.cpp because
it is very similar to some code already there, doing similar work,
and so that I can reuse a few helpers there without making them global.

When `this` is referenced, implicitly or explicitly, in a task construct
other than for a field access, it currently refers directly to the `this`
formal, just as it does outside a task construct. Specifying a task intent
explicitly for `this` or for an individual field is currently unsupported.