Improve printing of some types
Resolves `test/types/sync/vass/sync-type-1.future` .

* Print sync, single types ex. as `sync int(64)` .
* Print managed non-nilable class types without `!` .
* When a generic type is instantiated with managed/sync/single types,
  print the user-readable form of the latter.
* No changes in the developer mode, however.
  Staying with our tradition of printing types differently
  when `--devel` is on vs. off.

#### Implementation Note

There is subtle timing interplay between:

* AggregateType::getNewInstantiation()
* TypeSymbol::renameInstantiatedSingle()
* fixTypeNames(), which calls toString(Type*,bool)

which I do not claim to understand. Seems like the ideal place
to set the type's name is in renameInstantiatedSingle() and
getting rid of fixTypeNames(). However this may be too early
because nested types may not have been resolved yet.
I mitigate some of this by moving the call to renameInstantiatedSingle()
to a later spot in getNewInstantiation().

I am removing `test/types/sync/vass/sync-type-1.chpl` altogether
because printing of sync types is exercised in a dozen other tests.