Fix extern functions with ref return intent
There was an (untested) error about extern functions not being
able to be setters, but this error does not make sense any more.

Long ago, functions with `var` return intent were "setters", and
that caused them to create 2 instantiations of the function, with
`param setter` set differently depending on how they were used. These
have been replaced with return-intent-overloading. As a result,
the `ref` return intent no longer has special behavior inside of
the function body and so can apply to C functions.