Add regression related to passing `nil` to extern functions
This test worked a few days ago (e.g., as of tag
1.20.0-alpha-2019-08-20), but is not working on master now.  Since
then we've started inserting casts of nil to the expected integer
pointer type, but C doesn't like these casts.  This could probably be
fixed in the long-term by preserving type alias names through to
codegen time (a long-time good intention), but the short-term fix
would probably be to not cast 'nil's passed to extern pointer
arguments.