Mason: mason-publish takes a registry as an argument  (#13703)
Mason publish now accepts a registry as a positional argument instead of only publishing to the official mason-registry. You can pass a git path for a remote registry or a local path to a local registry. For local registries, publish tags and commits your changes so if MASON_REGISTRY points to both registries you can use your local package immediately after `mason publish <path-to-local-registry>`. Throws errors when there is an issue with the registry you target, or the package you are trying to publish. 


-Changed noFork.chpl and the .good to more accurately test with the new changes
-Changed publish.chpl to test with new additions
-Changed the help tests to match the new help outputs
-Changed errors to be throw MasonErrors, added call to update the registrys before we publish.
-Changed dryTest.good output to match new thrown error
-Changed packageName.chpl test to catch the thrown error it generates
-Changed publish.good, and both help tests to match new errors that are thrown
-Updated MasonHelp output to include '--no-update' argument
-Added thrown MasonErrors when package already exists in bricks with that version.
-Added two tests to check when a local path to a registry is a bad local registry.