Better nilability error message upon returns, yields
Before this change, the error message upon assigning to a non-nilable
owned/shared from its nilable counterpart came from the corresponding
assignment operator in the modules.

Since we implement returns and yields with assignment, the similar
error upon returning/yielding also came from that assignment operator.
Therefore the phrasing of the error talked about assignment.

This change adds where-clauses to these assignment operators so that
no operator is applicable when in the above error scenarios.
This way, the error is generated directly by the compiler.
The compiler already adjusts the phrasing of the error message upon
an unresolved call (assignment, in the above scenarios) to indicate
assignment, return, or yield as appropriate.

So, in these illegal returning and yielding scenarios, the text
of the error message now talks about returning or yielding,
not about assignment, as illustrated by the newly-added tests.

A couple .good files for errors upon assignments needed adjustment
because of different error text between the modules and the compiler.