Enable --no-legacy-classes on internal modules
Resolves https://github.com/Cray/chapel-private/issues/350.

#13447 applied the class type changes only to user code to get started.
Enabling it also for internal and standard modules.

In my experience adjusting the modules, it was sometimes difficult
to figure out what needs adjustment. For example, to figure out that
the management decoration was missing on `parentHandle`, I cloned
the code then was reducing it step by step until about 50 lines of
decorations, when I finally saw it. Another example was the missing
decoration in `var nodes : [{0..#nSlots}] (int, int, LocalDequeNode(eltType)?);`
Also, none of the error pointed me directly to the missing decorations
on `DistDeque._impl` and in the return expression for `DistDeque._value`.

Looking through the module code, I feels that `unmanaged` is more common
than we expected when making the default decoration be `borrowed`.

I was surprised to observe that this declaration:
```chpl
var rootLocale : locale? = nil;
```
inferred the `borrowed` decoration on `rootLocale` instead of issuing a
"generic management" error.

When adding a management decoration on `var x: SomeClass`, I used
`unmanaged` where I saw `x` being `delete`-ed manually.