Refactor "magic" of finding the body of a catch-with-test case
This refactors some of the magic code that I added to
checkResolved.cpp into a new method on CheckStmt called
bodyWithoutTest() which for a catch statement of the form: `catch e:
MyError { ... } returns the block containing the `{ ... }` block.

I'd hoped to also use this method in the errorHandling.cpp code that
contains similar logic for finding the conditional, but it wants to
manipulate the else-clause as well, so that didn't end up working out
as I'd hoped.

It's not clear to me whether there's a reason that we couldn't just
nest the catch statements to begin with such that:

```
try {
  ...
} catch (e: Err1) {
  ...
} catch (e: Err2) {
  ...
} catch {
  ...
}
```

couldn't just be stored as:

```
try {
  ...
}
CatchStmt
  if (e: Err1) then
    ...
  else
    CatchStmt
      if (e: Err2) then
        ...
      else
        CatchStmt
          ...
```

to more directly reflect the control flow, but exploring that was
more than I was able to take on at this point.