Make minimum string size a config param and 0-by-default (#13860)
This PR moves `private param chpl_string_min_alloc_size` from ByteBufferHelpers 
to ChapelBase as `config param chpl_stringMinAllocSize`.
 
Prior to string/bytes refactor (#13709), this was a param that was used in subset of 
string operations that allocates buffers. The functions that used this param were 
the ones that are about retrieving a byte or codepoint from a string.

After the refactor, all the allocations (and reallocations) come from ByteBufferHelpers,
which used this minimum size always. This caused tests that leak small strings leak 
even larger amounts.

It seemed like a better approach to make this a config param that is 0 by default. 
There may be cases where a lot of appends to smaller strings happen, in which case 
setting this to something reasonable can help users get rid of some of the memory 
allocations.

I have run some perf tests. It might help with some of the performance regressions. 
But I am not convinced that this'll be enough to bring us back to where we were in 
terms of performance.

Test:
- [x] standard
- [x] gasnet

[Review by @ronawho]