Add core mechanics for partial instantiations
This commit updates numerous files as a starting point for supporting
partial instantiations.

AggregageType.cpp is the most heavily impacted and contains a wide variety
of changes enabling ::generateType and its downstream calls to handle
out-of-order instantiation of fields. Special attention is made to maintain
the 'genericFields' list when new AggregateType* instances are created. As a
result of this new functionality, some error messages are no longer applicable
and have been removed.

A partial instantiation can occur when a type constructor call either includes
the '?' symbol or when the user did not provide an argument per field. The '?'
symbol indicates that fields with default values should not be evaluated unless
explicitly specified.

This commit also includes some support for nested partial instantiations,
primarily to allow for partial owned/shared expressions. The potential
use-cases for this feature are not entirely well understood, and will require
further consideration going forward.

The old way of updating a type's name assumed in-order instantiation of fields.
This commit removes the old implementation from TypeSymbol and adds new
functions in AggregateType.cpp that account for out-of-order instantiation.
Out-of-order instantiations will rely on named-expressions to accurately
represent the type to the user.

The representation of '?' is changed from a DefExpr to a new symbol called
'gUninstantiated'. This representation is easier to reason about in the AST
compared to a DefExpr of some arbitrary "chpl__query###" symbol. The parser is
updated to generate SymExprs to this new symbol. The parser is also updated to
allow for comparisons against '?', like "MyType.field == ?", in order to test
whether a field is instantiated. Unfortunately I was not able to find a way to
allow '?' on the left-hand side of the equality operators.

A new primitive, PRIM_IS_BOUND, is used to query whether a field is
instantiated and will support some Reflection functions.