Update compiler to handle gUninstantiated
This commit updates normalize() to handle partial instantiation representations
in the AST. The first step of this change is to handle gUninstantiated as a
SymExpr and not a DefExpr. This requires some additional branches and arguments
to functions, but the end result is largely the same. New errors are issued if
multiple '?' appear in a type constructor call (to discourage usage as a
positional argument). Tuples are excepted from these errors as otherwise we
could not indicate a fully-generic tuple (e.g. "(?,?,?)").

Other parts of normalize are updated to recognize a partial type expression
and to generate where-clause content as we would normally do for a generic
formal's type expression.

Other parts of resolution are updated to handle gUninstantiated or its type.
'expandVarArgs' also needs to differentiate between a DefExpr and a '?'
SymExpr. 'postFold' is updated to handle '?' in param expressions. 'preFold'
is updated to handle comparisons between expressions and '?' and to handle
the implementation of PRIM_IS_BOUND.