Change allowed intents for string arguments in multilocale libraries
Allow `in` and `const in` intents for strings in multilocale libraries, and ban
anything vaguely ref-like that involves communication back and forth.

While here, also update the error message for the single locale case - more than
just the listed ones are allowed, so it doesn't make sense to only say one of
the allowed ones in the error message.

I debated about blanket allowing more intents for single locale, but ultimately
decided against it - some of them run into an internal check which was intended
to ensure only strings were used but they got tangled up in it due to the
argument type being ref(string) at the time.  Resolving that should be do-able
but was more effort than I wanted to tackle this close to the release with other
priorities.