This makes in-module self-references work, but breaks module M storing M
Prior to this change, code like:

module M {
  proc foo() {
  }
  M.foo();
}

would break because we no longer put 'M' into the global scope, so
lookups of it would fail without a 'use' even though it seems silly to
require each module definition to 'use' itself.  Here, I put 'M'
within M's scope itself. so that it can be found.  However, this has
the downside of not permitting 'M' to contain a declaration of another
symbol named 'M' (for better or worse).