Added new check flag to mason publish (#13876)
Mason publish has a new flag `--check`. Runs a more complete set of checks and tests to assess a packages viability to be published. Either ran with no positional <registry> argument, which runs `--check` in relation to the official mason-registry, or with a personal registry path with `mason publish --check <path-to-registry>`.

Added tests within test/mason/publish-tests/.
Added --check to the help output.