Deprecate adding to associative domain by assigning to array
Add a warning when an associative array grows its domain by assigning to an
unused element.

Add a const accessor for `map`, used in some tests once they switch to `map`.

Update the TOML module to avoid doing this by using `map` instead.

Update mason to use the `map` in TOML instead of the associative domain.

Update a handful of tests to avoid this pattern.

This is a checkpoint so I can run testing and see what is left.