Updating documentation for mason.rst (#13937)
Updating documentation for mason.rst

Added all the documentation for `mason publish` For both publishing to the mason-registry or to a person registry, remote or local. 
Documented `--check` flag in relation to publishing as well as gave examples how to use local registries in conjunction with new commands. 

Added MASON_OFFLINE to the environment variable part of the docs.