Updates for array modifying domain deprecation
Update mason to use map in places where it was updating an associative domain
through an array. And to get it working with the updated TOML library.

Update tests to use maps instead of updating associative domains through
arrays.

Update the spec to avoid talking about the deprecated feature and remove a
test that used it.