Lock down intents for 1-D arrays in python interop
Similar to the case for c_strings and strings, we copy arrays from Python before
passing them to Chapel, so allowing intents other than `in` or `const in` is
incorrect.

While here, realized that using `userString` for the function would not work if
the wrapped function happened to have both an array argument and a string
argument - instead, just print the function name.