Fix precedence of !, owned, etc.
Before this change, the parser was treating `!` as the unary-`!` operator.
It had higher precedence than + and lower precedence than : .

This change moves TBANG in the precedence table. To keep the logical "not"
working as before, we mark one of its uses with a special `%prec` directive.

This also marks TBORROWED etc. as right associative.
This makes `borrowed borrowed X` parse as borrowed (borrowed X).
Which might not matter in practice.

Meanwhile postfix-`!` and postfix-`?` are left associative.
`x ! !` and `y ? ?` parse as `(x!)!` and `(y?)?` respectively.

Several tests are added to show these precedences in action.

The tests `type-casting.chpl` and `nilable-equalities.chpl` are adjusted
to reflect that `!` now binds more tightly than `owned` et al.