Fix to avoid reimplementing the wheel and encountering issues for library export
We were seeing an issue with multilocale runs because the internal and standard
modules define exported functions with c_string arguments and those would
ordinarily trigger the warning message.  However, we never generate mli
marshalling code for internal and standard module exported functions, since they
are intended for use solely within the runtime.  The solution is to extend the
check used by python interoperability and the header generation to be used by
mli and by the checks at resolution, so that we don't implement the same check
in three different places.