Update TestSystem.rst for multilocale C tests
In the section mentioning `.test.c` files, also mention `.ml-test.c` files for
multilocale testing.  In the list of file types, change the `.test.c`
description to mention it is for single locale and add a `.ml-test.c` line.

While here, also switch to using "multilocale" instead of "multi-locale" since
that seems to be the norm.