Primer for list operations (#14048)
Primer for list operations (#14048)

This PR introduces a primer for list operations. I tried to use an
informal, conversational style when writing this primer, as I felt
this matched the tone of tutorials that most helped me when I was
learning to program.

The scenario constructed in the primer is slightly contrived (what
hypothetical scenario isn't?), however I thought it better to tell a
story instead of just explaining list methods by rote.

Thanks to @lydia-duncan and @ben-albrecht for review.