Clean up the module file generation a little bit.
For the modulefile, instead of doing a 3-part dance with here-documents
respectively not subject to replacement, subject to replacement, and not
subject to replacement, use a single here-document and let sed do the
one replacement we need.

Stop insisting on finding a /etc/opt/cray/release/cray-release file
containing OS release information (which isn't available on Shasta login
nodes), just go by the CRAYPE_NETWORK_TARGET environment setting.  We
may need to revisit this later, if it turns out this isn't sufficiently
unique, but it's certainly good enough for now.

While here, get rid of the 'accel' setting, which wasn't used.