Add debugging output to mli marshalling routines about sizes
Prior to this change, we would just inform about the type that was being pushed
or pulled in the marshalling routines.  While this was useful, it was not
sufficient for determining what happened when, say, you didn't get all of your
string.

This change adds debugging output to say the length of the data being passed and
received.  It currently just fires for c_strings and strings (but that's all it
needs to fire for, at the moment).