Elminate multi-locale leaks in DimensionalDist2D
Switch the dimension specifiers from classes to records.
This simplifies their memory management and reduces the malloc/free overhead.
As a side effect, this eliminates leaks when using DimensionalDist2D
on multiple locales, when privatization kicks in.

This changes the public interface of a dimension specifier:
* removes `dsiSupportsPrivatization1d()`
* always requires `dsiGetPrivatizeData1d()` and `dsiPrivatize1d()`
  - the former can simply return `this`
  - the latter is now a type method, otherwise we'd have to copy
    the original record across locales, which defeats the purpose
    of dsiGetPrivatizeData1d()
  - future work: maybe we can simply copy the original record,
    without bothering with dsiGetPrivatizeData1d and dsiPrivatize1d ?

An interesting, albeit expected, change that I made to support the above
is to replace some variables from `const` (containing a class reference)
to `var` (containing a record).

While dimensional specifies are now created using `new` without `unmanaged`,
DimensionalDist2D itself is still a class, as per DSI, and so continues
to be "unmanaged". We do not delete it explicitly because it is wrapped in
a `dmap` aka `_distribution`, which deletes it once it is no longer needed.