Add public use of IO to CSV module and one of its tests
The CSVIO record takes a channel as an argument in one of its initializers, so
should publicly use IO so that channels are available.

I also added a use of the module to the test file, even though it wasn't
necessary.  This was mostly because I opened that file first, but it makes use
of other IO symbols so it might as well have a use in the interest of
documenting the code.