Enable lifetime constrains with varargs
Resolves #13509.

Allow the varargs formal to participate in lifetime constrains. For example:

    proc Test.dependsOn(tests...)  lifetime this < tests  {...}

Such a constraint means a conjunction of like constraints where the varargs
formal is replaced with each expanded formal in turn. For example, when
the above method is invoked with two actuals, the lifetime constraint means

    lifetime this < tests(1), this < tests(2)

Note that currently the compiler does not support this syntax in user code.

While this change was motivated by the lifetime clause in dependsOn()
in modules/packages/UnitTest.chpl, see #13509 for details, this change
actually removes that lifetime clause. The reason is that now that it works,
it generates lifetime errors, while we do not see anything incorrect.
These lifetime errors are independent of the varargs issue, that is,
they are also reported if we replace the varargs method with fixed-args
methods.

The arguments to dependsOn are first-class functions, so these errors
may be due to their lifetimes being not well defined or implemented.
This is something we need to work out once we properly define first-class
functions themselves.