Update the tests in the test/types/file directory to use IO
These tests were made to specifically test the IO module, so need an explicit
use of the module now that it is not included by default.

Some of these tests exercised failure modes, and so needed their expected line
numbers updated as a result.  One test also had an out of date comment, so I
removed it as well while I was there.