Fix assignment on 'owned'
Resolves #14147.

The assignment from owned to owned now succeeds when the l.h.s. is nilable
and the r.h.s. is `nil` of a nilable type.

It also produces a better error message at run time when the r.h.s. is `nil`
of a non-nilable type. Which can happen due to holes in static checking.

BTW here is a small-ish reproducer for #14147:
```chpl
proc create(gonil:bool): owned object? {
  if gonil then
    return nil;
  else
    return new owned object();
}

proc main() {
  var inst: owned object? = create(false);

  writeln("a ", inst);
  inst = create(true);  // used to halt here
  writeln("b ", inst);
}
```