Support fixed-heap-only scalable registration.
Even if the provider supports scalable registration we might want to
register only the registered heap.  (Say, because the hardware or the
provider cannot deal with page existence changes between registration
and use.)  Add support for that here.

While here, clean up the registered memory accessibility and add a
couple of header comments.