Compile LCALS without --legacy-classes
This includes a workaround for #13721 by using `LCALS_Overlapping_Array_3D`
instead of `LCALS_Overlapping_Array_3D(real)` as a formal argument type.
The lost `real` is checked with an assertion in the function body.
This latter can probably be left out safely, though.

When #13721 is resolved, this change to LCALSLoops.chpl can be reverted.