Update memory registration.
Memory registration support had some rough edges and bugs.

Get rid of mrGetLocalDesc() and just use mrGetDesc() instead.  Memory
region descriptors are by definition local -- they're specifically for
the case where the provider asserts FI_MR_LOCAL requiring registered
memory for local references.

In a few spots where we need to make sure that remote nodes can reach
local addresses, get the key for the virtual address in question rather
than the descriptor, since what we're asking about is remote reference.

For remote memory accesses, along with the key we also need to retrieve
the offset from the start of the memory region and use that as the
remote "address", rather than using the referenced virtual address
directly.  The latter would have worked as long as we were registering
the entire virtual address space with scalable registration, because the
VA is precisely the offset from 0, but not if we're only registering
part of it, such as for example a fixed heap.