Fix memory leak caused by use of first class functions (#14163)
Fix memory leak caused by use of first class functions (#14163)

This PR fixes a memory leak caused by the use of first class
functions. In the past, the wrapper that instantiates an instance of
a FCF returned an unmanaged class instance.

This PR resumes efforts started by @benharsh, and makes the FCF class
instance shared instead of unmanaged. This triggers class deletion
automatically when the last reference to a FCF instance goes out of
scope, closing the memory leaks.

Some of the issues mentioned by Ben related to `writeThis` for FCFs
seem to no longer be relevant (or at least, I did not run into
issues).

To fix test failures related to user specified FCF return types, I
altered the code that maps FCF return types, so that it returns a
resolved handle to `_shared(parent)` instead of `parent` (where
parent is the base class for a given FCF signature).

I had to adjust the .good file for only one test.

---

Testing:

- [x] `ALL` on `linux64` when `CHPL_COMM=none`
- [x] `ALL` on `linux64` when `CHPL_COMM=gasnet`

---

Thanks to @benharsh for review!