Fix annotations for compiler performance graphs
Compiler performance graphs use a couple of features that were breaking
annotation support. Dygraphs requires that annotations are "attached" to
a series, so in order for annotations to display we have to attach them
to a visible series. That was broken in 2 ways.

In genGraphs we build a single graph that contains all the compiler
passes and then from that graph we build another one that is only the
top ten passes. When picking the top 10 our graphkeys got out of date,
so we were trying to attach annotations to invalid series. Fix that by
pulling the annotation series directly from the generated json instead
of relying on the outdated graphkeys.

In perfgraph.js we expand the graph with all compiler passes into
multiple graphs with one pass each. When doing this we were still
attaching annotations to an old series. Fix that by deep cloning the
annotations and attaching to the series for the pass being displayed