Add use of SysError to the Error handling examples in the spec
The IO module has a public use of SysError.  With the IO module no longer used
by default, the SysError module is also no longer used by default.  This means
that an explicit use of the SysError module is necessary to reference the
errors it defines.

These examples call a function that throws.  However, their catch blocks look
for an error subclass which it does not throw, so it would not be appropriate
for the example defining the function to use SysError.  So add the use to the
example programs that reference the FileNotFoundError.