Update submitted shootouts for need for IO to be explicitly used
This was triggered by updating to master, as the tests were previously failing
earlier in compilation.  Since these are the submitted shootouts, we need to
wait until they are updated on the site to update them here.

Also, added .bad files for the tests that I futurized, so that we ensure the
failure mode is tracked by the .future file