More memory management updates, this time for XC/gni.
At some point in this series of commits I broke operation with the gni
provider on Cray XC systems.  This gets at least the memory management
part back.  The main change here is that we now support just two overall
registration modes: either the provider supports scalable registration
of the entire address space, or the user has to specify a fixed heap in
which case we will register that and only that.  (With the gni provider
we'll do the latter.)  While here add FI_MR_VIRT_ADDR support, in which
a remote address within a memory region really is a virtual address, not
the offset from the start of the region.

Things are still broken after this on XC/gni, but that seems to have to
do with CQ attributes and/or flags rather than memory.