Just use a single address vector.
A long time ago we tried to bind each of the multiple transmit contexts
of a scalable transmit endpoint individually to our address vector, and
got a -FI_EINVAL return value from the bind call.  This led to a belief
that we couldn't bind the same AV to multiple endpoints.  Then when the
code for non-scalable EPs was added, one of the things we thought we had
to do was to create a separate AV for each endpoint.

It turns out this isn't true.  Binding the same AV to multiple endpoints
is fine.  The reason we got the -FI_EINVAL return value when binding the
AV to the transmit contexts of our scalable transmit endpoint was that
in effect we were trying to bind the same AV to the same endpoint (the
scalable one) multiple times.  The transmit contexts don't even need to
have an AV bound to them; they just use the one bound to their endpoint.

Thus, here, recode to just use a single AV and bind it to all the
endpoints.