Correct the flags for binding completion counters to transmit endpoints.
We've been using the same flags when binding completion counters to
transmit endpoints (and contexts) as we do when binding completion
queues to them.  But in fact counters use rather different flags,
although it seems at least some providers are insensitive to that.
Here, use the correct flags.  This gets the gni provider working on Cray
XC systems again, although we still have a problem at shutdown because
it says our waitset is busy (has things still using it).  We'll address
that separately.