Simplify deciding whether or not to use a waitset.
Instead of relying on a user-set environment variable to tell us not to
use a waitset, simply use one unless the provider doesn't support them.
Except: don't use one with the gni provider, because although we can
create one there and it seems to function properly, when we go to close
it we get -FI_EBUSY indicating that something is still associated with
it.

We can look into the problem with the gni provider in the future if we
want to, but it's not pressing because using a waitset is a performance
optimization and when we want performance on XC we use comm=ugni rather
than comm=ofi.