Do PMI2 allgathers larger than the maximum KVS value length in chunks.
PMI2 defines a maximum key value length for KVS put/get.  In some cases
we allgather more than that.  Here, arrange to chunk the allgather in
that case.

It would be better to get the maximum length from an actual PMI-2
include file, but we don't currently have a trustworthy way to find such
an include file in an automated way during the runtime build, so that
will have to wait.