Clean up specifying the provider.
The user can specify the provider, if desired, using either the Chapel
CHPL_RT_COMM_OFI_PROVIDER environment variable or the built-in libfabric
FI_PROVIDER one, with the former overriding the latter.  Up until now
the override has been accomplished by putting the value of the Chapel
env var into the provider field of the hints structure passed to the
fi_getinfo() call.  This works, but it isn't actually a documented
behavior; while the man pages describe both FI_PROVIDER and the hints
prov_name field as being ways to specify the provider, they don't say
which wins out when both are set.

Here, instead of putting the CHPL_RT_COMM_OFI_PROVIDER value into the
hints prov_name field, use it to set the FI_PROVIDER environment
variable, overriding any value already there.  This makes the override
explicit, while also avoiding the weirdness of needing to malloc() the
hints_prov_name field because fi_freeinfo() will free() it, which isn't
very well documented.

While here, get rid of the setting of FI_LOCAL_MR in the hints 'mode'
field when we're going to be using a Cray XC Aries network.  This is
superfluous now that we set FI_MR_LOCAL in the hints mr_mode field, and
according to the man page is a technique that should only be used with
older libfabric releases (v1.4 and earlier) anyway.

Also while here, rename some variables in isInProviderName() to avoid
a case of identifier shadowing which might be confusing to readers.